requirements.txt
discord.py[voice]
sqlalchemy
flask
flask_wtf
